:art: test
test


Former-commit-id: 98c239221fdb6b06d63635e26c8aa8eeca8fb992

+++ BE/blur/blur-auth/src/main/java/com/blur/auth/api/entity/User.java
@@ -1,5 +1,7 @@
 package com.blur.auth.api.entity;
 
+import java.util.UUID;
+
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
@@ -33,7 +35,7 @@ public class User {
     @Id
     @Column(name = "user_no")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer userNo;
+    private Long userNo;
 
     @Column(name = "user_id", length = 64, unique = true)
     @NotNull
